requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
PyJWT==2.9.0
bcrypt==4.2.1
python-multipart==0.0.20
